Fixed some bugs and added network name in the DAO

+++ cache_scripts/aragon/collectors/vote.py
@@ -46,8 +46,15 @@ def update_votes(meta_data: Dict, net: str, endpoints: Dict) -> None:
 
     filename: str = os.path.join('datawarehouse', 'aragon', f'{META_KEY}.csv')
 
-    # Always rewrite the whole file cause it is more efficient to do it than request all the open proposals.
-    df.to_csv(filename, index=False)
+    if not df.empty:
+        if os.path.isfile(filename):
+            # Always rewrite the whole file cause it is more efficient to do it than request all the open proposals.
+            dff = pd.read_csv(filename, header=0)
+            dff = [dff['network'] == net]
+            df = df.append(dff)
+            df.to_csv(filename, index=False)
+        else:
+            df.to_csv(filename)
 
     print(f'Data stored in {filename}.\n')
